+++ audionimbus/tests/hrtf.rs
@@ -1,0 +1,9 @@
+#[test]
+fn test_load_hrtf_sofa_filename() {
+    // TODO: implement test.
+}
+
+#[test]
+fn test_load_hrtf_sofa_buffer() {
+    // TODO: implement test.
+}